+++ src/main/java/com/ecsa/divisas/config/postgresConfig.java
@@ -1,7 +1,0 @@
-package com.ecsa.divisas.config;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "spring.datasource")
-public class postgresConfig {
-}