+++ src/main/java/com/ecsa/divisas/repository/divisaRepository.java
@@ -1,7 +1,0 @@
-package com.ecsa.divisas.repository;
-
-import com.ecsa.divisas.model.divisa;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface divisaRepository extends JpaRepository<divisa, Long> {
-}