+++ src/main/java/com/ecsa/divisas/repository/transactionRepository.java
@@ -1,10 +1,9 @@
 package com.ecsa.divisas.repository;
 
-import com.ecsa.divisas.model.currencyModel;
+import com.ecsa.divisas.model.transactionModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface currencyRepository extends JpaRepository<currencyModel, Long> {
-
+public interface transactionRepository extends JpaRepository<transactionModel, Long> {
 }