+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 #V 1.0.0
 Modulo Forex
  * Creacion de Compras
+ * Modificacion tipo BD SQLite